+++ src/stores/ActionEnum.js
@@ -4,7 +4,7 @@
     REMOVE_TASK : "remove-task",
     CHANGE_TASK : "change-task",
     TOGGLE_TASK : "toggle-task",
-    SET_OVERDUE : "set-overdue"
+    COUNT_DAYS_LEFT : "set-overdue"
 }
 
 export default ActionEnum
