+++ src/utils/index.js
@@ -1,4 +1,3 @@
 export  {loadFromLocalStorage}  from "./loadFromLocalStorage";
 export  {saveToLocalStorage} from "./saveToLocalStorage";
 export {StorageEnum} from "./StorageEnum"
-export {CheckDeadline} from "./checkDeadline"
