+++ src/utils/loadFromLocalStorage.js
@@ -1,5 +1,5 @@
 
-export function loadFromLocalStorage ({key: key}) {
+export function loadFromLocalStorage ({key}) {
  const data = localStorage.getItem(key);
  if (!data) return "";
  return JSON.parse(data)
